+++ src/test/java/tech/deplant/java4ever/frtest/unit/datatype/UintTests.java
@@ -1,0 +1,4 @@
+package tech.deplant.java4ever.frtest.unit.datatype;
+
+public class UintTests {
+}